Piece movement shown
Pieces are now updated after the end of every cycle; each pit has the correct number of piece representations inside it now! woo! Current colors of pieces are placeholders
Still working on piece representation inside the stores, though.
Also ironed out a lot of problems with matchup between the GUI's and gamemanager's versions of the pits. Still having trouble capturing, probably the issue with that is in the link between GUI and backend.