Server-Client Updated
I've successfully ran a full game of mancala on this, (putting almost all the inputs through the server program). The thing for whose turn it currently is was hacked a little bit to get this running.

Also fixed most of the condition checks in handle_input(): they're .equals() (which works) rather than == (which doesn't)

Also moved updating display to... clickign the stores, for now. This is a workaround until I can figure out how to get the threads to play nicely such that the buffer handlers can call that in some way.

Fixed pit hover-over
Removed some debugging text