Time, Illegal moves now cause loss
As specs require, illegal and time-out moves now result in failure.

GUI does not allow inputting illegal moves (since that seems frustrating and awful for the user)

Handling of win/loss conditions, or at least a stopgap of that, is also done. They're pretty simple for now.

handle_input() now returns a string, which is then used by the buffer handler to perform other tasks.